+++ lib/extensions.dart
@@ -1,0 +1,6 @@
+// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
+
+/// Exports the extension methods exposed by this package.
+library extensions;
+
+export 'src/term_similarity_extensions.dart';