+++ lib/src/_index.dart
@@ -2,7 +2,8 @@
 // End user is granted a non-exclusive non-transferable license the ("License")
 // to use GM Consult's proprietary software (the "Software").
 
-export 'term_comparison_extensions.dart';
+export 'term_similarity.dart';
+export 'term_similarity_extensions.dart';
 export 'text_tokenizer.dart';
 export 'text_document.dart';
 export 'token.dart';