+++ lib/type_definitions.dart
@@ -1,0 +1,6 @@
+// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
+
+/// Exports all the typedefs used in this package.
+library type_definitions;
+
+export 'src/typedefs.dart';