+++ test/text_analysis_test.dart
@@ -7,58 +7,17 @@ import 'package:text_analysis/src/_index.dart';
 import 'package:test/test.dart';
 import 'data/sample_news.dart';
 import 'data/sample_stocks.dart';
+import 'data/data.dart';
 
 void main() {
   group('TextTokenizer', () {
     //
 
-    /// Three paragraphs of text used for testing tokenization.
-    ///
-    /// Includes numbers, currencies, abbreviations, hyphens and identifiers
-    final text = [
-      'The Dow Jones rallied even as U.S. troops were put on alert amid '
-          'the Ukraine crisis. Tesla stock fought back while Apple '
-          'stock struggled. ',
-      '[TSLA.XNGS] Tesla\'s #TeslaMotor Stock Is Getting Hammered.',
-      'Among the best EV stocks to buy and watch, Tesla '
-          '(TSLA.XNGS) is pulling back from new highs after a failed breakout '
-          'above a \$1,201.05 double-bottom entry. ',
-      'Meanwhile, Peloton reportedly finds an activist investor from the Cote D\'Azure knocking '
-          'on its door.  In a scathing new letter released Monday, activist '
-          'Tesla Capital is pushing for Peloton to fire CEO, Chairman and '
-          'founder John Foley and explore a sale.'
-          'The stock crash has fueled by strong  suspicions of mismanagement.'
-    ];
-
-    final json = {
-      'avatarImageUrl':
-          'https://firebasestorage.googleapis.com/v0/b/buysellhold-322d1.appspot.com/o/logos%2FTSLA%3AXNGS.png?alt=media&token=c365db47-9482-4237-9267-82f72854d161',
-      'description': 'A stock split gave a nice short-term boost to Amazon Inc'
-          'in late May and in early June. Alphabet has a planned '
-          'stock split for next month. Tesla is also waiting on '
-          'shareholder approval for a 3-for-1 stock split. ',
-      'entityType': 'NewsItem',
-      'hashTags': ['#Tesla'],
-      'id': 'ee1760a1-a259-50dc-b11d-8baf34d7d1c5',
-      'itemGuid':
-          'trading-shopify-stock-ahead-of-10-for-1-stock-split-technical-analysis-june-2022?puc=yahoo&cm_ven=YAHOO&yptr=yahoo',
-      'linkUrl':
-          'https://www.thestreet.com/investing/trading-shopify-stock-ahead-of-10-for-1-stock-split-technical-analysis-june-2022?puc=yahoo&cm_ven=YAHOO&yptr=yahoo',
-      'locale': 'Locale.en_US',
-      'name': 'Shopify Stock Split What the Charts Say Ahead of 10-for-1 Split',
-      'publicationDate': '2022-06-28T17:44:00.000Z',
-      'publisher': {
-        'linkUrl': 'http://www.thestreet.com/',
-        'title': 'TheStreet com'
-      },
-      'timestamp': 1656464362162
-    };
-
     test('TextDocument', (() async {
       // Initialize a StringBuffer to hold the source text
       final sourceBuilder = StringBuffer();
       // Concatenate the elements of [text] using line-endings
-      for (final src in text) {
+      for (final src in TextAnalysisTestData.text) {
         sourceBuilder.writeln(src);
       }
       // convert the StringBuffer to a String
@@ -80,73 +39,11 @@ void main() {
       print('Flesch-Kincaid Grade Level: ${textDoc.fleschKincaidGradeLevel()}');
     }));
 
-    test('jaccardSimilarityMap', (() {
-      final k = 2;
-      final term = 'broder';
-      final candidates = ['board', 'broad', 'boardroom', 'border'];
-      print('k-grams with k = $k');
-      final map = term.jaccardSimilarityMap(candidates, k);
-
-      for (final other in map.entries) {
-        print(
-            'Similarity between $term and ${other.key} = ${other.value.toStringAsFixed(4)}');
-      }
-    }));
-
-    test('String extension `matches`', (() {
-      final k = 2;
-      final term = 'bodrer';
-      final candidates = [
-        'board',
-        'broad',
-        'boarder',
-        'border',
-        'brother',
-        'bored'
-      ];
-      final matches = term.matches(candidates, k: k);
-      print(matches);
-    }));
-
-    test('lengthSimilarity', (() {
-      final term = '123456789012';
-      print('-'.padRight(26, '-'));
-      print('${'this'.padRight(4)}:'
-          '${'other'.padLeft(6)}'
-          '${'ld'.padLeft(8)}'
-          '${'ls'.padLeft(8)}');
-      print('-'.padRight(26, '-'));
-      for (var i = 0; i < 48; i++) {
-        final other = ''.padRight(i, '@');
-        final ld = term.lengthDistance(other);
-        final ls = term.lengthSimilarity(other);
-        print('${term.length.toString().padRight(4)}:'
-            '${i.toString().padLeft(6)}'
-            '${ld.toStringAsFixed(2).padLeft(8)}'
-            '${ls.toStringAsFixed(2).padLeft(8)}');
-      }
-      print('-'.padRight(26, '-'));
-    }));
-
-    test('kGram(k)', () async {
-      final source = text.first;
-      // use a TextTokenizer instance to tokenize the source
-      final tokens = await TextTokenizer().tokenize(source);
-      // get the bi-grams
-      final Map<String, Set<Term>> kGramIndex = tokens.kGrams(3);
-      // add the tri-grams
-      // kGramIndex.addAll(document.tokens.kGrams(3));
-      print('${'k-gram'.padRight(8)} Terms Set');
-      for (final entry in kGramIndex.entries) {
-        print('${entry.key.padRight(8)} ${entry.value}');
-      }
-    });
-
     test('TextTokenizer.tokenize', () async {
       // Initialize a StringBuffer to hold the source text
       final sourceBuilder = StringBuffer();
       // Concatenate the elements of [text] using line-endings
-      for (final src in text) {
+      for (final src in TextAnalysisTestData.text) {
         sourceBuilder.writeln(src);
       }
       // convert the StringBuffer to a String
@@ -163,7 +60,7 @@ void main() {
       // Initialize a StringBuffer to hold the source text
       final sourceBuilder = StringBuffer();
       // Concatenate the elements of [text] using line-endings
-      for (final src in text) {
+      for (final src in TextAnalysisTestData.text) {
         sourceBuilder.writeln(src);
       }
       // convert the StringBuffer to a String
@@ -191,8 +88,8 @@ void main() {
 
     test('TextTokenizer.tokenizeJson', () async {
       // use a TextTokenizer instance to tokenize the json with 3 zones
-      final tokens = await TextTokenizer()
-          .tokenizeJson(json, ['name', 'description', 'hashTags']);
+      final tokens = await TextTokenizer().tokenizeJson(
+          TextAnalysisTestData.json, ['name', 'description', 'hashTags']);
       // map the document's tokens to a list of terms (strings)
       final terms = tokens.map((e) => e.term).toList();
       // print the terms
@@ -202,7 +99,8 @@ void main() {
     /// Tokenize JSON with NO zones.
     test('TextTokenizer.tokenizeJson', () async {
       // use a TextTokenizer instance to tokenize the json with NO zones
-      final tokens = await TextTokenizer().tokenizeJson(json);
+      final tokens =
+          await TextTokenizer().tokenizeJson(TextAnalysisTestData.json);
       // map the document's tokens to a list of terms (strings)
       final terms = tokens.map((e) => e.term).toList();
       // print the terms
@@ -247,19 +145,4 @@ void main() {
       print('Found ${entries.length} terms');
     });
   });
-
-  group('k-grams', (() {
-    //
-    test('jaccardSimilarity', (() {
-      final k = 2;
-      final term = 'bord';
-      final candidates = ['board', 'broad', 'boardroom', 'border'];
-      print('k-grams with k = $k');
-      for (final other in candidates) {
-        final jCf = term.jaccardSimilarity(other, k);
-        print(
-            'Similarity between $term and $other = ${jCf.toStringAsFixed(4)}');
-      }
-    }));
-  }));
 }